requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
azure-core==1.30.2
azure-identity==1.16.1
backports.zoneinfo==0.2.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
Deprecated==1.2.14
exceptiongroup==1.2.1
frozenlist==1.4.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
importlib-metadata==7.1.0
importlib-resources==6.4.0
microsoft-kiota-abstractions==1.3.3
microsoft-kiota-authentication-azure==1.0.0
microsoft-kiota-http==1.3.1
microsoft-kiota-serialization-form==0.1.0
microsoft-kiota-serialization-json==1.2.0
microsoft-kiota-serialization-multipart==0.1.0
microsoft-kiota-serialization-text==1.0.0
msal==1.28.1
msal-extensions==1.1.0
msgraph-core==1.0.0
msgraph-sdk==1.4.0
multidict==6.0.5
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
packaging==24.1
pendulum==3.0.0
portalocker==2.8.2
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
sniffio==1.3.1
std-uritemplate==1.0.1
time-machine==2.14.1
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
